Create Description.txt
Artificial Neural Networks for Building Energy Modeling. If you want to learn more about this project, please read my thesis file. Otherwise, these are the questions that this projects answers.

Is it possible to build an accurate Neural Network-Based Surrogate Model from a physics-based model?

Can we use a Genetic Algorithm to optimize building energy consumption of Surrogate Models?

What are the benefits of the Genetic Algorithm compared to the Bayesian Algorithm when optimizing surrogate models?